requirements.txt
argparse
jsonlib-python3
pyOpenSSL
datetime
dnspython
requests
httpx